+++ Airship/AirshipCore/Tests/SubscriptionListAPIClientTest.swift
@@ -12,7 +12,7 @@ class SubscriptionListAPIClientTest: XCTestCase {
     var client: SubscriptionListAPIClient!
 
     override func setUpWithError() throws {
-        self.config = RuntimeConfig(config: Config(), dataStore: PreferenceDataStore(keyPrefix: UUID().uuidString))
+        self.config = RuntimeConfig(config: Config(), dataStore: PreferenceDataStore(appKey: UUID().uuidString))
         self.session = TestRequestSession.init()
         self.client = SubscriptionListAPIClient(config: self.config, session: self.session)
     }